movement-strategy: alter the 'next parallel clue' algorithm
This new algorithm determines which next parallel word shares the most
adjacent fillable boxes with the current word, and uses that word.  In
the case where multiple parallel words share the same amount of adjacent
boxes, then the word closest to the start of the current word will be
used.