Add support for periodically downloading puzzles in the background.
This is implemented by using JobScheduler to periodically download
puzzles that the user has selected. JobScheduler is only available
on API version >=21, so  this functionality is disabled for lower
API versions.

Using the JobScheduler API allows us to easily
provide options for downloading on wifi only, when charging, etc.

Currently just attempt to download every hour. There are more
intelligent things we can do here, but this has been working well
on my device.

The updated NYT authentication mechanism is supported by posting a
notification requesting login when a background download fails
instead of immediately launching the login activity.

A slight refactor of Downloaders was required to allow for
downloading the most recent the most recent puzzle for each source
instead of just the puzzle for the current (or specified) date.