Add display scratch option
Notes are displayed on playboard in small letters if the box is
otherwise blank.  Needed to add to Box.java an across and down clue
number and position calculated during setBoxes.  Additionally, the
PlayboardRenderer draw functions now have two arguments:
displayScratchAcross and displayScratchDown to indicate whether to add
notes to the board.  The drawBox function fetches and renders the
scratch notes if required.  Two options are needed because it is not
pleasing to show the scratch on the notes page for the clue being edited
(but you do want to see the scratch for other clues).

An option has been added to preferences to enable/disable this
behaviour.  Default off.