Fix performance of SeparatedRecyclerViewAdapter.
When creating the ViewHolder for a puzzle item, the adapter was iterating over every section in the adapter, inflating a layout, and keeping only the last one.

I believe that this causes the time required to render the main BrowseActivity to grow quadratically with the number of sections in the BrowseActivity.

For a large crossword library (several years worth) on a Pixel 2, this decreases the time to open the BrowseActivity and have a usable UI from over a minute to several seconds.  The time now seems to be dominated by time required to perform the file system operations to list the puzzles.