Movement strategy settings fix.
When you change this setting it's not reflected until the app is restarted
which initially made me think it wasn't working. Additionally the up/back arrow
in settings does nothing so I've made that just go back.

Possibly the onResume/onStart code could be refactored a bit more. I'd have a go
if you wanted.